Book,Copy Endpoints Restructerd, added more test cases , validators
