+++ src/endpoints/book/__init__.py
@@ -1,0 +1,6 @@
+from src.endpoints.book.delete.book_delete import book_delete
+from src.endpoints.book.get.get_book_by_id import get_book_by_id
+from src.endpoints.book.get.get_books_by_query import get_books_by_query
+from src.endpoints.book.post.book_create import book_create
+from src.endpoints.book.put.book_update import book_update
+from src.endpoints.book.router_init import router
