+++ src/endpoints/book/router_init.py
@@ -1,0 +1,5 @@
+from fastapi import APIRouter
+
+router = APIRouter(
+    prefix="/book", tags=["book"], responses={401: {"book": "Book not Found"}}
+)
