+++ src/endpoints/copy/__init__.py
@@ -1,0 +1,7 @@
+from src.endpoints.copy.delete.copy_delete import copy_delete
+from src.endpoints.copy.get.get_copies import get_copies
+from src.endpoints.copy.get.get_copies_by_book_id import get_copies_by_book_id
+from src.endpoints.copy.get.get_copy_by_id import get_copy_by_id
+from src.endpoints.copy.post.copy_create import copy_create
+from src.endpoints.copy.put.copy_update import copy_update
+from src.endpoints.copy.router_init import router
