+++ src/endpoints/copy/router_init.py
@@ -1,0 +1,5 @@
+from fastapi import APIRouter
+
+router = APIRouter(
+    prefix="/copy", tags=["copy"], responses={401: {"user": "Not authorized"}}
+)
